complete some TODOs
